+++ src/pch.h
@@ -1,17 +1,23 @@
 #pragma once
 
-#include <SKSE/Impl/PCH.h>
 #include <RE/Skyrim.h>
+#include <SKSE/Impl/PCH.h>
 #include <SKSE/SKSE.h>
 
 #include <Windows.h>
+#include <bcrypt.h>
 
 #include <algorithm>
 #include <array>
 #include <atomic>
 #include <cstddef>
 #include <cstdint>
+#include <filesystem>
+#include <limits>
+#include <span>
+#include <string>
 #include <string_view>
+#include <vector>
 
 namespace logger = SKSE::log;
 using namespace std::literals;